[nexus] Change the design of the MMU
Extend the MMU to 16 registers and just use each of them to describe the
physical memory page (8 bits) of each page frame.

That is much simpler than before, and allow the programmer to map any
frame to any memory page in the whole address space. What was a huge
limiter before this.